perf: mmap blob reads + BufWriter ingestion (v1.1.1)
Phase D/E/G: u32 record IDs, Arc<str> string interning via StringPool,
minor session/index cleanups (~2.75 GB savings for 5M events).

Phase F: offload requestParameters/responseElements/additionalEventData
to a temp file via BlobStore, keeping only 12-byte BlobRef in RAM
(~1.8 GB savings). Fixed post-Phase-F perf regression:
- BufWriter during ingestion batches ~15M blob writes into large flushes
- seal() after ingestion mmaps the blob file (memmap2); all subsequent
  reads are lock-free pointer arithmetic — restores pre-F filter speed

All 107 core tests passing.

+++ crates/core/src/detection/mod.rs
@@ -31,7 +31,7 @@ pub struct Alert {
     /// True count of matching records (may exceed matching_record_ids.len()).
     pub matching_count: usize,
     /// Up to 100 matching record IDs (capped for IPC efficiency).
-    pub matching_record_ids: Vec<u64>,
+    pub matching_record_ids: Vec<u32>,
     pub metadata: HashMap<String, String>,
     pub mitre_tactic: String,
     pub mitre_technique: String,
